Moving a lot of business logic from controller and DAO into the service

+++ app/org/sagebionetworks/bridge/dynamodb/DynamoSurveyDao.java
@@ -103,21 +103,18 @@ public class DynamoSurveyDao implements SurveyDao {
             return surveyMapper.queryPage(DynamoSurvey.class, query).getCount();
         }
         
-        List<Survey> getAll(boolean exceptionIfEmpty) {
+        List<Survey> getAll() {
             List<DynamoSurvey> dynamoSurveys = null;
             if (surveyGuid == null) {
                 dynamoSurveys = queryBySecondaryIndex();
             } else {
                 dynamoSurveys = query();
             }
-            if (exceptionIfEmpty && dynamoSurveys.size() == 0) {
-                throw new EntityNotFoundException(DynamoSurvey.class);
-            }
             return ImmutableList.copyOf(dynamoSurveys);
         }
         
-        Survey getOne(boolean exceptionIfEmpty) {
-            List<Survey> surveys = getAll(exceptionIfEmpty);
+        Survey getOne() {
+            List<Survey> surveys = getAll();
             if (!surveys.isEmpty()) {
                 if (!skipElements) {
                     attachSurveyElements(surveys.get(0));    
@@ -260,9 +257,6 @@ public class DynamoSurveyDao implements SurveyDao {
 
     @Override
     public Survey publishSurvey(StudyIdentifier study, Survey survey, boolean newSchemaRev) {
-        if (survey.isDeleted()) {
-            throw new EntityNotFoundException(Survey.class);
-        }
         if (!survey.isPublished()) {
             // update survey
             survey.setPublished(true);
@@ -287,21 +281,6 @@ public class DynamoSurveyDao implements SurveyDao {
     public Survey updateSurvey(Survey survey) {
         Survey existing = getSurvey(survey, false);
         
-        boolean undeletion = existing.isDeleted() && !survey.isDeleted();
-        if (!undeletion && existing.isDeleted()) {
-            throw new EntityNotFoundException(Survey.class);
-        }
-        if (existing.isPublished()) {
-            // If the existing survey is published, the only thing you can do is undelete it.
-            if (undeletion) {
-                existing = getSurvey(survey, true); // get all the children for the update
-                existing.setDeleted(false);
-                return saveSurvey(existing);
-            } else {
-                throw new PublishedSurveyException(survey);
-            }
-        }
-
         // copy over mutable fields
         existing.setIdentifier(survey.getIdentifier());
         existing.setName(survey.getName());
@@ -322,9 +301,6 @@ public class DynamoSurveyDao implements SurveyDao {
     @Override
     public Survey versionSurvey(GuidCreatedOnVersionHolder keys) {
         DynamoSurvey existing = (DynamoSurvey)getSurvey(keys, true);
-        if (existing.isDeleted()) {
-            throw new EntityNotFoundException(Survey.class);
-        }        
         DynamoSurvey copy = new DynamoSurvey(existing);
         copy.setPublished(false);
         copy.setDeleted(false);
@@ -349,48 +325,47 @@ public class DynamoSurveyDao implements SurveyDao {
 
     @Override
     public void deleteSurveyPermanently(GuidCreatedOnVersionHolder keys) {
-        
-        Survey existing = new QueryBuilder().setSurvey(keys.getGuid()).setCreatedOn(keys.getCreatedOn())
-                .setSkipElements(true).getOne(true);
-        deleteAllElements(existing.getGuid(), existing.getCreatedOn());
-        surveyMapper.delete(existing);
-        
-        // Delete the schemas as well, or they accumulate.
-        try {
-            StudyIdentifier studyId = new StudyIdentifierImpl(existing.getStudyIdentifier());
-            uploadSchemaService.deleteUploadSchemaById(studyId, existing.getIdentifier());
-        } catch(EntityNotFoundException e) {
-            // This is OK. Just means this survey wasn't published.
+        Survey existing = getSurvey(keys, false);
+        if (existing != null) {
+            deleteAllElements(existing.getGuid(), existing.getCreatedOn());
+            surveyMapper.delete(existing);
+            // Delete the schemas as well, or they accumulate.
+            try {
+                StudyIdentifier studyId = new StudyIdentifierImpl(existing.getStudyIdentifier());
+                uploadSchemaService.deleteUploadSchemaById(studyId, existing.getIdentifier());
+            } catch(EntityNotFoundException e) {
+                // This is OK. Just means this survey wasn't published.
+            }
         }
     }
 
     @Override
     public List<Survey> getSurveyAllVersions(StudyIdentifier studyIdentifier, String guid, boolean includeDeleted) {
-        return new QueryBuilder().setStudy(studyIdentifier).setSurvey(guid).setDeleted(includeDeleted).getAll(true);
+        return new QueryBuilder().setStudy(studyIdentifier).setSurvey(guid).setDeleted(includeDeleted).getAll();
     }
     
     @Override
     public Survey getSurveyMostRecentVersion(StudyIdentifier studyIdentifier, String guid) {
-        return new QueryBuilder().setStudy(studyIdentifier).setSurvey(guid).setDeleted(false).getOne(true);
+        return new QueryBuilder().setStudy(studyIdentifier).setSurvey(guid).setDeleted(false).getOne();
     }
 
     @Override
     public Survey getSurveyMostRecentlyPublishedVersion(StudyIdentifier studyIdentifier, String guid, boolean includeElements) {
         return new QueryBuilder().setStudy(studyIdentifier).isPublished().setSurvey(guid).setDeleted(false)
-                .setSkipElements(!includeElements).getOne(true);
+                .setSkipElements(!includeElements).getOne();
     }
     
     // secondary index query (not survey GUID) 
     @Override
     public List<Survey> getAllSurveysMostRecentlyPublishedVersion(StudyIdentifier studyIdentifier, boolean includeDeleted) {
-        List<Survey> surveys = new QueryBuilder().setStudy(studyIdentifier).isPublished().setDeleted(includeDeleted).getAll(false);
+        List<Survey> surveys = new QueryBuilder().setStudy(studyIdentifier).isPublished().setDeleted(includeDeleted).getAll();
         return findMostRecentVersions(surveys);
     }
     
     // secondary index query (not survey GUID)
     @Override
     public List<Survey> getAllSurveysMostRecentVersion(StudyIdentifier studyIdentifier, boolean includeDeleted) {
-        List<Survey> surveys = new QueryBuilder().setStudy(studyIdentifier).setDeleted(includeDeleted).getAll(false);
+        List<Survey> surveys = new QueryBuilder().setStudy(studyIdentifier).setDeleted(includeDeleted).getAll();
         return findMostRecentVersions(surveys);
     }
     
@@ -402,7 +377,7 @@ public class DynamoSurveyDao implements SurveyDao {
     @Override
     public Survey getSurvey(GuidCreatedOnVersionHolder keys, boolean includeElements) {
         return new QueryBuilder().setSurvey(keys.getGuid()).setCreatedOn(keys.getCreatedOn())
-                .setSkipElements(!includeElements).getOne(true);
+                .setSkipElements(!includeElements).getOne();
     }
     
     /**
